% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_coefs.R
\name{compare_coefs}
\alias{compare_coefs}
\title{Compare how a coefficient differs between models}
\usage{
compare_coefs(m1, m2, coef)
}
\arguments{
\item{The}{input are the two model objects and a string for the coefficients name. This should be the same as displayed in the model summary.}
}
\value{
It returns a vector with the difference, z- and p-value.
}
\description{
This function compares the coefficient of one variable in different models.
It can be used for mediation analysis to check if a mediator significantly reduces the coefficient of that variable.
Another use case is when the model doesn't allow a specification of an interaction effect like the did package.
Here you can run the model for different groups and compare the estimates.
The function should only be used for comparing models with the same dependent variable unless it is standardized.
}
