% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_coefs.R
\name{compare_coefs}
\alias{compare_coefs}
\title{Compare how a coefficient differs between models}
\usage{
compare_coefs(m1, m2, coef)
}
\arguments{
\item{m1}{A model object. The first model to compare.}

\item{m2}{A model object. The second model to compare.}

\item{coef}{A string. The name of the coefficient to compare, as it appears in the model summary.}
}
\value{
A vector with the difference in the coefficient, the z-value, and the p-value.
}
\description{
This function compares the coefficient of one variable across two models.
It can be used for mediation analysis to check if a mediator significantly reduces the coefficient of that variable.
Another use case is when the model doesn't allow a specification of an interaction effect like the did package.
Here you can run the model for different groups and compare the estimates.
The function should only be used for comparing models with the same dependent variable unless it is standardized.
}
\details{
It is compatible with lm(), lm_robust(), feols() and aggte().
The formula (based on Paternoster et al. (1998)) for the calculation is:
\deqn{Z = \frac{b_1 - b_2}{\sqrt{SE_{b_1}^2 + SE_{b_2}^2}}}
}
\references{
Paternoster, R., Brame, R., Mazerolle, P., & Piquero, A. (1998). Using the Correct Statistical Test for the Equality of Regression Coefficients. \emph{Criminology, 36}(4), 859–866. https://doi.org/10.1111/j.1745-9125.1998.tb01268.x
}
